Enhancement: Extract StringMatchesFormatDescriptionTest
